Type the ghost import delay as int
The delay is interpolated into a relative date string, which accepts a
fraction without complaint and then reads something else entirely: the
sign and the whole part are dropped and the decimals become positive
hours, so '-2.5 hours' and '-0.5 hours' both mean +5, and '-1.25 hours'
means +25. That pushes the cutoff into the future, where dateModified
is below it for every import in progress, and the query that looks for
stuck imports would mark healthy ones as failed instead.

Whole hours are all it can express, and the one caller passes 2.

+++ app/bundles/LeadBundle/Entity/ImportRepository.php
@@ -13,9 +13,9 @@ class ImportRepository extends CommonRepository
     /**
      * Count how many imports with the status is there.
      *
-     * @param float $ghostDelay when is the import ghost? In hours
+     * @param int $ghostDelay when is the import ghost? In hours
      */
-    public function getGhostImports(float $ghostDelay = 2, ?int $limit = null): array
+    public function getGhostImports(int $ghostDelay = 2, ?int $limit = null): array
     {
         $q = $this->getQueryForStatuses([Import::IN_PROGRESS]);
         $q->select($this->getTableAlias())
